[#1197] Add "maxSlices" pie chart param to group slices into "Others"
A pie chart doesn't work well if there're many categories, so this code adds a
"maxSlices" parameter to it where, if there're more slices than this number,
it'll group them in a single "Others" slice.

We do this by getting the totals from the "summary" attribute, and subtracting
the values of all slices that will still be shown.

This works OK but there's a caveat: it'll break if we're using anything other
tham a simple sum. For example, if the data contains averages we'll display
incorrect data.

In my opinion, this is OK, as a pie chart isn't well suited to display averages
anyway. We just need to take care not to use averages (or other summary methods)
when limiting the number of slices.

opentrials/opentrials#1197