Fix bug with JQuery usage in PivotTable's Angular binding
Taking the example of using `babbage.ui` inside `os-viewer`, this is what
happens:

On `os-viewer`, we run `window.$ = window.jQuery = require('jquery')` to add
jQuery to the global scope.

On `babbage.ui`'s PivotTable, we run `import 'pivottable'`, which internally
runs `require('jquery')` and adds itself to the returned instance.

To allow multiple versions of the same dependency to coexist, Webpack returns
a different instance when you require jquery on `os-viewer` and on `babbage.ui`
(on `os-viewer` we use jQuery v2.1.4, while on `babbage.ui` we use v1.12.4
because of `bubbletree`). This means that if you require jQuery inside
`babbage.ui`, it'll return v1.12.4, while if you do the same inside
`os-viewer`, it'll return v2.1.4.

As we're running `import 'pivottable'` inside `babbage.ui`, the pivottable will
be added to jQuery v1.12.4. However, in the PivotTable's Angular binding, we're
using the global $ instance, which is set to `os-viewer`'s version.

All of this means that we use the wrong jQuery version inside the Angular
binding, which raises an exception because `pivottable` wasn't loaded in this
version. This commit fixes this by explictly requiring jQuery, instead of using
the global `$`. Webpack will make sure we're using the correct version.